+++ src/Test/Tasty/QuickCheck/Laws.hs
@@ -19,6 +19,7 @@ module Test.Tasty.QuickCheck.Laws (
   , module Test.Tasty.QuickCheck.Laws.ReaderMonad
   , module Test.Tasty.QuickCheck.Laws.StateMonad
   , module Test.Tasty.QuickCheck.Laws.WriterMonad
+  , module Test.Tasty.QuickCheck.Laws.MaybeMonad
 ) where
 
 import Test.Tasty.QuickCheck.Laws.Applicative
@@ -31,3 +32,4 @@ import Test.Tasty.QuickCheck.Laws.Monoid
 import Test.Tasty.QuickCheck.Laws.ReaderMonad
 import Test.Tasty.QuickCheck.Laws.StateMonad
 import Test.Tasty.QuickCheck.Laws.WriterMonad
+import Test.Tasty.QuickCheck.Laws.MaybeMonad