+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog for tasty-quickcheck-laws
 Unreleased changes
 ------------------
 
+* Added
+    * Maybe monad laws
 * Fixed
     * Clarified license (it's BSD3)
     * Bug in reference Reader monad implementation,